==> sfi-asset-level-data-develop/phase-2-deployment/src/main/jupyter/02-l8-tir-macro-localization-model-build/00-l8-tir-aws-management.ipynb <==
# View assets on S3

import boto3
import os
import shutil

! pip install awscli

s3 = boto3.resource('s3')
bucket = s3.Bucket('sfi-shared-assets')

! aws s3 ls s3://sfi-shared-assets/

## Landsat-8 TIR Macro-localization Model Build Chips

! aws s3 ls s3://sfi-shared-assets/L8-TIR-macro-localization-model-build/

! aws s3 ls s3://sfi-shared-assets/L8-TIR-macro-localization-model-build/L8-TIR-model-results/

! aws s3 ls s3://sfi-shared-assets/L8-TIR-macro-localization-model-build3/

! aws s3 ls s3://sfi-shared-assets/L8-TIR-macro-localization-model-build3/L8-TIR-model-results3/

## Remove from S3?

! aws s3 rm s3://sfi-shared-assets/L8-TIR-macro-localization-model-build3/L8-TIR-model-results3vgg13_multiclass_final.pkl

! aws s3 rm s3://sfi-shared-assets/L8-TIR-macro-localization-model-build/ALD_L8_TIR_steel_chips_v4p1_2017_train2.tar
