==> sfi-asset-level-data-develop/phase-2-deployment/src/main/jupyter/02-l8-tir-macro-localization-model-build/03.2-l8-tir-dl-model-data-prep-steel.ipynb <==
# Prepare Landsat-8 TIR chips for Macro-localization Deep Learning Model Build

This notebook prepares image chips to train a classifier on Landsat 8 TIR Band 10 image chips, and store them on AWS/S3.

## Import Libraries

import os
import shutil
from pathlib import Path

import boto3
import numpy as np
import pandas as pd

import rasterio
import sklearn.model_selection

## Download .tar Files From S3 Bucket and Extract Contents

CURRENT_DIRECTORY = os.getcwd()
AWS_SOURCE_PATH = 'L8-TIR-macro-localization-model-build3'

TARGET_PATH = '/scratch/ALD_L8_TIR_chips_v4p1_train3_steel_binary'

IMG_DIRS = (
    ('ALD_L8_TIR_landcover_chips_v4p1_2018_train3', 'landcover'),
    ('ALD_L8_TIR_cement_chips_v4p1_2018_train3', 'cement'),
    ('ALD_L8_TIR_cement_chips_v4p1_2019_train3', 'cement'),
    ('ALD_L8_TIR_cement_chips_v4p1_2020_train3', 'cement'),
    ('ALD_L8_TIR_steel_chips_v4p1_2017_train3', 'steel'),
    ('ALD_L8_TIR_steel_chips_v4p1_2018_train3', 'steel'),
    ('ALD_L8_TIR_steel_chips_v4p1_2019_train3', 'steel'),
    ('ALD_L8_TIR_steel_chips_v4p1_2020_train3', 'steel')
)

!mkdir -p {TARGET_PATH}

s3 = boto3.resource('s3')
bucket = s3.Bucket('sfi-shared-assets')

for source_file, _ in IMG_DIRS:
    bucket.download_file(str(Path(AWS_SOURCE_PATH, source_file+'.tar')), 
                         str(Path(TARGET_PATH, source_file + '.tar')))

for source_file, _ in IMG_DIRS:
    !cd {TARGET_PATH} && tar xf {str(Path(source_file + '.tar'))}

## Partition the Data Using Stratified Random Sampling

To help address the issue of limited sample sizes (in particular for steel plant imagery), we partitition the data using stratified random sampling.

* Will define PNGs to put in train/ and validate/ folders

image_list = ! find {TARGET_PATH} | grep tif$
class_assignments = [f.split('/')[-1].split('_')[1] for f in image_list]

train_idx, val_idx = next(sklearn.model_selection.StratifiedShuffleSplit(n_splits=2, random_state=42, test_size=0.2).split(class_assignments, class_assignments))
subset_assignments = ['train' if i in train_idx else 'validate' for i in range(len(image_list))]

output_pngs = [f.split('/')[-1].replace('tif', 'png') for f in image_list]

# replace cement with landcover
class_assignments = ['landcover' if assignment == 'cement' else assignment for assignment in class_assignments]

for image_class in np.unique(class_assignments):
    for subset in np.unique(subset_assignments):
        !mkdir -p {TARGET_PATH}/{subset}/{image_class}

## Convert GeoTiff to PNG (Includes Normalisation Step)

Fastai appears to require converting TIFF files to an alternative image format. Thus, convert from GeoTIFF to PNG.

def convert_image(tif_filename, png_filename):
    with rasterio.open(tif_filename) as infile:
        
        profile = infile.profile
        profile['driver'] = 'PNG'
        
        raster = infile.read()
        
        with rasterio.open(png_filename, 'w', **profile) as dst:
            dst.write(raster)

Convert each image only if its corresponding target file does not already exist.

for image_file, class_assignment, subset_assignment, png_file in zip(image_list, 
                                                                     class_assignments, 
                                                                     subset_assignments,
                                                                     output_pngs):
    if not Path(TARGET_PATH, subset_assignment, class_assignment, png_file).exists():
        convert_image(image_file, 
                      Path(TARGET_PATH, subset_assignment, class_assignment, png_file))

## Write out record of training/testing chips

train_record_pdf = pd.DataFrame({"file": output_pngs,
                                 "class": class_assignments,
                                 "subset": subset_assignments})

train_record_pdf.to_csv("../../resources/macro-loc-model-build/"+TARGET_PATH.split('/')[-1]+"_record.csv",
                       index=False)

## Tar Files and Upload to S3 Bucket

for source_file, _ in IMG_DIRS:
    shutil.rmtree(TARGET_PATH+'/'+source_file)
    os.remove(TARGET_PATH+'/'+source_file+'.tar')

unix_code = 'tar -C /scratch -cvf '+TARGET_PATH.split('/')[-1]+'.tar '+TARGET_PATH.split('/')[-1]
os.system(unix_code)

bucket.upload_file(TARGET_PATH.split('/')[-1]+'.tar', 
                   AWS_SOURCE_PATH+'/'+TARGET_PATH.split('/')[-1]+'.tar')

## Clean up Temporary Files

shutil.rmtree(TARGET_PATH)
os.remove(TARGET_PATH.split('/')[-1]+'.tar')
